+++ src/utils.h
@@ -38,13 +38,13 @@ namespace easyVDB
         for (int i = 0; i < hex.length(); i++) {
             if (hex[i] >= 48 && hex[i] <= 57)
             {
-                result += (hex[i] - 48) * pow(16, hex.length() - i - 1);
+                result += (hex[i] - 48) * static_cast<long long>(pow(16, hex.length() - i - 1));
             }
             else if (hex[i] >= 65 && hex[i] <= 70) {
-                result += (hex[i] - 55) * pow(16, hex.length() - i - 1);
+                result += (hex[i] - 55) * static_cast<long long>(pow(16, hex.length() - i - 1));
             }
             else if (hex[i] >= 97 && hex[i] <= 102) {
-                result += (hex[i] - 87) * pow(16, hex.length() - i - 1);
+                result += (hex[i] - 87) * static_cast<long long>(pow(16, hex.length() - i - 1));
             }
         }
         return result;
